Fix iOS 7.1 NSInteger string formatting warnings
-also #if 0 out unused newlyFullStationName NSString variable

+++ DockSmart/DockSmartAppDelegate.m
@@ -94,7 +94,7 @@ static NSString * const stationErrorMessage = @"Information might not be up-to-d
     UILocalNotification *triggeredNotification = [launchOptions objectForKey:UIApplicationLaunchOptionsLocalNotificationKey];
     
     
-    NSString* logText = [NSString stringWithFormat:@"willFinishLaunchingWithOptions: startLocation %@ triggeredNotification %@ applicationState: %d", startLocation, [triggeredNotification alertBody], [application applicationState]];
+    NSString* logText = [NSString stringWithFormat:@"willFinishLaunchingWithOptions: startLocation %@ triggeredNotification %@ applicationState: %ld", startLocation, [triggeredNotification alertBody], [application applicationState]];
     DLog(@"%@",logText);
     [[NSNotificationCenter defaultCenter] postNotificationName:kLogToTextViewNotif
                                                         object:self
@@ -109,7 +109,7 @@ static NSString * const stationErrorMessage = @"Information might not be up-to-d
  */
 - (void)application:(UIApplication *)application didReceiveLocalNotification:(UILocalNotification *)notification
 {
-    NSString* logText = [NSString stringWithFormat:@"applicationDidReceiveLocalNotification: applicationState: %d", [application applicationState]];
+    NSString* logText = [NSString stringWithFormat:@"applicationDidReceiveLocalNotification: applicationState: %ld", [application applicationState]];
     DLog(@"%@",logText);
     [[NSNotificationCenter defaultCenter] postNotificationName:kLogToTextViewNotif
                                                         object:self
@@ -151,7 +151,7 @@ static NSString * const stationErrorMessage = @"Information might not be up-to-d
 {
     // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
     // Use this method to pause ongoing tasks, disable timers, and throttle down OpenGL ES frame rates. Games should use this method to pause the game.
-    NSString* logText = [NSString stringWithFormat:@"applicationWillResignActive: applicationState: %d", [application applicationState]];
+    NSString* logText = [NSString stringWithFormat:@"applicationWillResignActive: applicationState: %ld", [application applicationState]];
     DLog(@"%@",logText);
     [[NSNotificationCenter defaultCenter] postNotificationName:kLogToTextViewNotif
                                                         object:self
@@ -165,7 +165,7 @@ static NSString * const stationErrorMessage = @"Information might not be up-to-d
     // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later. 
     // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
     
-    NSString* logText = [NSString stringWithFormat:@"applicationDidEnterBackground: applicationState: %d", [application applicationState]];
+    NSString* logText = [NSString stringWithFormat:@"applicationDidEnterBackground: applicationState: %ld", [application applicationState]];
     DLog(@"%@",logText);
     [[NSNotificationCenter defaultCenter] postNotificationName:kLogToTextViewNotif
                                                         object:self
@@ -184,7 +184,7 @@ static NSString * const stationErrorMessage = @"Information might not be up-to-d
 {
     // Called as part of the transition from the background to the inactive state; here you can undo many of the changes made on entering the background.
 
-    NSString* logText = [NSString stringWithFormat:@"applicationWillEnterForeground: applicationState: %d", [application applicationState]];
+    NSString* logText = [NSString stringWithFormat:@"applicationWillEnterForeground: applicationState: %ld", [application applicationState]];
     DLog(@"%@",logText);
     [[NSNotificationCenter defaultCenter] postNotificationName:kLogToTextViewNotif
                                                         object:self
@@ -200,7 +200,7 @@ static NSString * const stationErrorMessage = @"Information might not be up-to-d
 {
     // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
     
-    NSString* logText = [NSString stringWithFormat:@"applicationDidBecomeActive: applicationState: %d", [application applicationState]];
+    NSString* logText = [NSString stringWithFormat:@"applicationDidBecomeActive: applicationState: %ld", [application applicationState]];
     DLog(@"%@",logText);
     [[NSNotificationCenter defaultCenter] postNotificationName:kLogToTextViewNotif
                                                         object:self
@@ -223,7 +223,7 @@ static NSString * const stationErrorMessage = @"Information might not be up-to-d
 - (void)applicationWillTerminate:(UIApplication *)application
 {
     // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
-    NSString* logText = [NSString stringWithFormat:@"applicationWillTerminate: applicationState: %d", [application applicationState]];
+    NSString* logText = [NSString stringWithFormat:@"applicationWillTerminate: applicationState: %ld", [application applicationState]];
     DLog(@"%@",logText);
     [[NSNotificationCenter defaultCenter] postNotificationName:kLogToTextViewNotif
                                                         object:self
@@ -466,7 +466,7 @@ static NSString * const stationErrorMessage = @"Information might not be up-to-d
         else
         {
             //otherwise just log an error and go to the next station
-            DLog("Invalid station coordinate: stationID %d, %@, coord %f, %f", tempStation.stationID, tempStation.name, tempStation.latitude, tempStation.longitude);
+            DLog("Invalid station coordinate: stationID %ld, %@, coord %f, %f", (long)tempStation.stationID, tempStation.name, tempStation.latitude, tempStation.longitude);
         }
     }
